feat(ui): ui updates 2
