+++ app/src/main/java/com/kev/instantsystem/ui/InstantSystemApp.kt
@@ -13,8 +13,8 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import com.kev.instantsystem.ui.discovery.DiscoveryRoute
 import com.kev.instantsystem.ui.home.HomeRoute
+import com.kev.instantsystem.ui.headlines.HeadlinesRoute
 import com.kev.instantsystem.ui.menu.MenuScreen
 import com.kev.instantsystem.ui.navigation.BottomNavItem
 import com.kev.instantsystem.ui.navigation.MainBottomBar
@@ -82,10 +82,14 @@ private fun ISNavHost(
             )
         }
         composable(BottomNavItem.Search.route) {
-            SearchRoute()
+            SearchRoute(
+                onDetailVisibilityChanged = onDetailVisibilityChanged
+            )
         }
-        composable(BottomNavItem.Discovery.route) {
-            DiscoveryRoute()
+        composable(BottomNavItem.Headlines.route) {
+            HeadlinesRoute(
+                onDetailVisibilityChanged = onDetailVisibilityChanged
+            )
         }
         composable(BottomNavItem.Bookmarks.route) {
             BookmarksScreen()
