+++ app/src/main/java/com/kev/instantsystem/ui/components/NewsListScreenPaging.kt
@@ -1,13 +1,11 @@
-package com.kev.instantsystem.ui.home
+package com.kev.instantsystem.ui.components
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import androidx.paging.LoadState
 import androidx.paging.compose.LazyPagingItems
