+++ app/src/main/java/com/kev/instantsystem/ui/home/HomeListPane.kt
@@ -1,4 +1,4 @@
-package com.kev.instantsystem.ui.discovery
+package com.kev.instantsystem.ui.home
 
 import androidx.compose.runtime.Composable
 import androidx.paging.LoadState
@@ -7,10 +7,10 @@ import com.kev.instantsystem.domain.model.Article
 import com.kev.instantsystem.ui.components.EmptyListPlaceholder
 import com.kev.instantsystem.ui.components.LoaderScreen
 import com.kev.instantsystem.ui.components.LoaderState
-import com.kev.instantsystem.ui.home.NewsListScreenPaging
+import com.kev.instantsystem.ui.components.NewsListScreenPaging
 
 @Composable
-fun DiscoveryListPane(
+fun HomeListPane(
     articles: LazyPagingItems<Article>,
     onArticleClick: (Article) -> Unit
 ) {
