+++ app/src/main/java/com/kev/instantsystem/ui/home/HomeViewModel.kt
@@ -5,34 +5,25 @@ import androidx.lifecycle.viewModelScope
 import androidx.paging.PagingData
 import androidx.paging.cachedIn
 import com.kev.instantsystem.domain.model.Article
-import com.kev.instantsystem.domain.usecase.GetTopHeadlinesUseCase
+import com.kev.instantsystem.domain.usecase.SearchEverythingUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
+/**
+ * ViewModel responsible for managing the state and logic of the Journal screen.
+ * Fetches paginated news articles using the "everything" endpoint.
+ */
 @HiltViewModel
 class HomeViewModel @Inject constructor(
-    private val getTopHeadlinesUseCase: GetTopHeadlinesUseCase
+    private val searchEverythingUseCase: SearchEverythingUseCase
 ) : ViewModel() {
 
-    val categories = listOf(
-        ArticleCategory.Latest,
-        ArticleCategory.Science,
-        ArticleCategory.Sports,
-        ArticleCategory.Technology,
-        ArticleCategory.Health,
-        ArticleCategory.Business
-    )
-
-    val articlesMap: Map<ArticleCategory, Flow<PagingData<Article>>> =
-        categories.associateWith { getArticles(it.category) }
-
-    private fun getArticles(category: String?): Flow<PagingData<Article>> {
-        return getTopHeadlinesUseCase(
-            country = "us",
-            category = category,
-            query = null
-        ).cachedIn(viewModelScope)
-    }
-
+    /**
+     * Public stream of articles based on a fixed or dynamic search.
+     * For now, we provide a default query (e.g., "technology").
+     */
+    val discoveryArticles: Flow<PagingData<Article>> = searchEverythingUseCase(
+        query = "technology" // you can make this dynamic later
+    ).cachedIn(viewModelScope)
 }
