+++ app/src/main/java/com/kev/instantsystem/ui/navigation/NavigationRoutes.kt
@@ -4,6 +4,6 @@ object NavigationRoutes {
     const val HOME = "home"
     const val SEARCH = "search"
     const val BOOKMARK = "bookmark"
-    const val DISCOVERY = "discovery"
+    const val HEADLINES = "headlines"
     const val MENU = "menu"
 }
